Incorperated into main
Incorperated mysql into main, it now creates the database as well as its tables when starting up for the first time without the mysql database. SInce it runs on create if not exists, then each time F1_GO is started up, it will simply see that its already there and just continue. This allowes for updates to the mysql down the road to be seamlessly intigrated when rolling out to users.